Add flush to release
Some upper-layer filesystems such as NFS never (or almost never) issue
flush calls under normal circumstances.  This causes files created on
s3fs under nfsd to never be written and always show up as zero-sized.

This patch adds a flush call to the release method in order that the
data be written.